backend update2

+++ src/main/java/hello/service/question_codeService.java
@@ -7,7 +7,7 @@ import hello.model.question_code;
 
 public interface question_codeService {
 	 void save(question_code code);
-	 int findbyid(question_code mq);
+	 long findbyid(question_code mq);
 	 void updatecode(question_code mq);
 	 void delete(question_code mq);
 }
